pyproject.toml
[project]
name = "extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { email = "ben.epstein97@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "baml-py==0.81.3",
    "cloudpathlib>=0.21.0",
    "deltalake>=0.25.5",
    "duckdb>=1.2.1",
    "fastapi[standard]>=0.115.11",
    "httpx>=0.28.1",
    "modal==0.73.115",
    "pdf2image>=1.17.0",
    "polars>=1.26.0",
    "prefect>=3.3.3",
    "prefect-gcp>=0.6.5",
    "prefect-github>=0.3.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "coverage >=7.4.0, <8",
    "mypy >=1.10, <2",
    "pytest >=8.2, <9",
    "pytest-cov >=5.0.0, <6",
    "ruff >=0.4.5, <1.0",
    "jupyter",
    "notebook",
    "ipython",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [ "src/extractor", "src/flows", "src/modals", "src/shared"]


[tool.mypy]
check_untyped_defs = true
exclude = ["baml_src/", "baml_client/", "shared/server.py"]


[[tool.mypy.overrides]]
module = "ruamel.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "prefect_gcp.*"
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true


[tool.ruff]
line-length = 120

[tool.ruff.lint]
# Include rules for: pyFlakes, pycodestyle (Err/Warn), pyDocstyle, Isort, Flake8-FIXme, PyLinT, NumPY, debugger
# See rules and codes: https://docs.astral.sh/ruff/rules/
select = ["F", "E", "W", "D", "I001", "FIX", "PL", "NPY", "T100"]
dummy-variable-rgx = "_$|dummy"
ignore = [
    "FIX002",   # block TODO comments
    "D104",     # require package __init__ docstring
    "D100",     # require module file docstring
    "D101",     # Class docstrings
    "D107",     # Module docstrings
    "D103",     # function docstrings
    "D102",     # public method docstrings
    "PLR2004",  # Magic value in comparison
]
extend-safe-fixes = ["F401"]

[tool.ruff.lint.per-file-ignores]
"src/shared/baml_client/**" = ["D205"]
"tests/**" = ["PLR2004"]

[tool.ruff.lint.pylint]
max-args = 10


[tool.ruff.lint.pydocstyle]
convention = "google"